pyproject.toml
[project]
name = "img-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "William Keating", email = "54231777+keatingw@users.noreply.github.com" },
]
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20.0",
    "datasets>=3.2.0",
    "fastapi[standard]>=0.115.6",
    "pillow>=11.1.0",
    "pydantic-settings>=2.7.1",
    "pydotenv>=0.0.7",
    "torch>=2.5.1",
]

[project.scripts]
img-search-api = "img_search.api.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["ipython>=8.31.0", "mypy>=1.14.0", "ruff>=0.8.4"]

[tool.ruff]
fix = true                                                  # autofix on (only for fixable list)
show-fixes = true                                           # show fixes made
include = ["*.py", "*.pyi", "**/pyproject.toml", "*.ipynb"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
pydocstyle.convention = "google"
select = ["ALL"]
ignore = [
    "E501",    # ignore line length
    "PD013",   # allow stack as well as melt
    "PD901",   # allow 'df' dataframe naming
    "PD010",   # allow pandas .pivot method
    "PD009",   # `iat` is allowed
    "PD011",   # allow `.values` methods given false positives
    "B023",    # allow loops to set new variables
    "B905",    # allow non-specific zip strictness
    "PLR2004", # allow literals in comparisons
    "D105",    # don't force docstrings on magic methods
    "D107",    # don't force docstrings on init
    "FIX002",  # allow TODO
    "RUF012",  # ignore mutable classvar
    "RUF013",  # implicit optional handled by mypy strict
    "COM812",  # allow no trailing comma
    "TD003",   # no issue links on todos
    "G004",    # allow fstring logging
    "SLF001",  # allow private access so lxml doesn't complain
    "ANN401",  # allow Any on kwargs
    "ISC001",  # conflict with formatter
    "CPY001",  # allow no copyright notices
]
fixable = [
    "I001",   # autofix for isort
    "RUF022", # autofix for __all__
]

[tool.ruff.lint.per-file-ignores]
"**/tests/*.py" = ["S101", "INP001", "D104"]

[tool.mypy]
plugins = ["pydantic.mypy"]
strict = true
disallow_any_unimported = true
python_executable = "./.venv/bin/python"

[[tool.mypy.overrides]]
module = ["datasets.*"]
ignore_missing_imports = true
